requirements.txt
Flask_WTF==0.14.2
SQLAlchemy==1.3.12
numpy==1.18.1
Flask_SQLAlchemy==2.4.1
matplotlib==3.1.2
redis==3.3.11
Flask_Script==2.0.6
Flask==1.1.1
Werkzeug==0.16.0
Flask_Migrate==2.5.2
alembic==1.3.3
Flask_Session==0.3.1
django_recaptcha==2.0.6
Pillow==7.0.0
